% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticModelTesting.R
\name{concordance}
\alias{concordance}
\title{Concordance}
\usage{
concordance(df, target, probability)
}
\arguments{
\item{df}{dataframe which target and variable present.}

\item{target}{Target variable.}

\item{probability}{Predicted probabilities.}
}
\value{
Concordance outputs
}
\description{
Checking Concordance.
}
\examples{
data("CTDF")
mylogit<-glm(TARGET~HOLDING_PERIOD+NO_OF_L_CR_TXNS+AGE_BKT,data=CTDF,family = "binomial")
summary(mylogit)
vif(mylogit)
CTDF$prob<-predict.glm(mylogit,CTDF,type="response")
concordance(CTDF,"TARGET","prob")
}
