% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticModelTesting.R
\name{hosmerlem_gof}
\alias{hosmerlem_gof}
\title{Chi Sq - Goodness of Fit}
\usage{
hosmerlem_gof(df, target, probability, g = 10)
}
\arguments{
\item{df}{dataframe which target and variable present.}

\item{target}{Target variable.}

\item{probability}{Predicted probabilities.}
}
\value{
Chi Sq. values
}
\description{
Checking Chi Sq - Goodness of Fit.
}
\examples{
data("CTDF")
mylogit<-glm(TARGET~HOLDING_PERIOD+NO_OF_L_CR_TXNS+AGE_BKT,data=CTDF,family = "binomial")
summary(mylogit)
vif(mylogit)
CTDF$prob<-predict.glm(mylogit,CTDF,type="response")
hosmerlem_gof(CTDF,"TARGET","prob")
}
\author{
Peter D. M. Macdonald, McMaster University
}
