% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticModelTesting.R
\name{KS}
\alias{KS}
\title{Ks statistics & ROC curve}
\usage{
KS(df, target, probability)
}
\arguments{
\item{df}{dataframe which target and variable present.}

\item{target}{Target variable.}

\item{probability}{Predicted probabilities.}
}
\value{
ROC curve and KS value
}
\description{
Plot ROC curve and calculate KS statistics.
}
\examples{
data("CTDF")
mylogit<-glm(TARGET~HOLDING_PERIOD+NO_OF_L_CR_TXNS+AGE_BKT,data=CTDF,family = "binomial")
summary(mylogit)
vif(mylogit)
CTDF$prob<-predict.glm(mylogit,CTDF,type="response")
KS(CTDF,"TARGET","prob")
}
\author{
Rajesh Jakhotia
}
