% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticModelTesting.R
\name{ROTable}
\alias{ROTable}
\title{Rank Ordering}
\usage{
ROTable(df, target, probability)
}
\arguments{
\item{df}{dataframe which target and variable present.}

\item{target}{Target variable.}

\item{probability}{Predicted probabilities.}
}
\value{
Rank ordering table
}
\description{
Produce Rank ordering table
}
\examples{

data("CTDF")
mylogit<-glm(TARGET~HOLDING_PERIOD+NO_OF_L_CR_TXNS+AGE_BKT,data=CTDF,family = "binomial")
summary(mylogit)
vif(mylogit)
CTDF$prob<-predict.glm(mylogit,CTDF,type="response")
ROTable(CTDF,"TARGET","prob")
}
\author{
Rajesh Jakhotia
}
