% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticModelTesting.R
\name{Rrate}
\alias{Rrate}
\title{Rrate function}
\usage{
Rrate(df, target, var)
}
\arguments{
\item{x}{varibale need to look the target distribution}
}
\value{
It will give back you the deciled data with target disribution
}
\description{
Takes the data and divid in to 10 parts and looks the target distribution.
}
\examples{
data("CTDF")
Rrate(df = CTDF,target = "TARGET",var = "HOLDING_PERIOD")
}
\author{
Rajesh Jakhotia
}
